requirements.txt
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.20.0
trimesh>=3.20.0
openai>=1.0.0
anthropic>=0.15.0
requests>=2.28.0
flask>=2.3.0
flask-cors>=3.0.0
PyQt6>=6.0.0
pyopengl>=3.1.0
pyqtgraph>=0.12.0
python-dotenv>=0.19.0
onnxruntime>=1.14.0
onnx>=1.12.0
pyinstaller>=4.0.0
aiohttp>=3.7.0
websockets>=10.0
scipy>=1.9.0
scikit-learn>=1.2.0
huggingface-hub>=0.15.0
tqdm>=4.60.0
